+++ src/GRA.Domain.Repository/IPrizeWinnerRepository.cs
@@ -9,5 +9,7 @@ namespace GRA.Domain.Repository
         Task<ICollection<PrizeWinner>> PageByWinnerAsync(int siteId, int userId, int skip, int take);
         Task<int> CountByWinningUserId(int siteId, int userId, bool? redeemed = null);
         Task<PrizeWinner> GetUserTriggerPrizeAsync(int userId, int triggerId);
+        Task<ICollection<PrizeWinner>> GetRedemptionsAsync(ReportCriterion criterion);
+        Task<ICollection<PrizeWinner>> GetUserPrizesAsync(ReportCriterion criterion);
     }
 }